requirements.txt
python-telegram-bot==21.3
groq
python-dotenv==1.0.1
google-generativeai==0.7.2
asyncpg